+++ rollup.config.js
@@ -6,10 +6,9 @@ import commonjs from 'rollup-plugin-commonjs'
 import json from 'rollup-plugin-json'
 import nodeResolve from 'rollup-plugin-node-resolve'
 import postcss from 'rollup-plugin-postcss'
+import { terser } from 'rollup-plugin-terser'
 import typescriptPlugin from 'rollup-plugin-typescript'
-import { uglify } from 'rollup-plugin-uglify'
 import typescript from 'typescript'
-import { minify } from 'uglify-es'
 import pkg from './package.json'
 
 const plugins = [
@@ -32,7 +31,7 @@ const plugins = [
       cssnano({ preset: 'default' }),
     ],
   }),
-  !process.env.ROLLUP_WATCH && uglify({}, minify),
+  !process.env.ROLLUP_WATCH && terser(),
 ]
 
 export default [