+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Support HTML whitelisting ([#26](https://github.com/marp-team/marp-core/pull/26))
+- Use [rollup-plugin-terser](https://github.com/TrySound/rollup-plugin-terser) instead of rollup-plugin-uglify ([#27](https://github.com/marp-team/marp-core/pull/27))
 
 ## v0.0.4 - 2018-08-29